// src/views/wallet/arbitrum/contract/contracts/interfaces/v1/IEzTreasury.ts
export type SwapQuoteStruct = {
  sellToken: PromiseOrValue<string>;
  buyToken: PromiseOrValue<string>;
  sellAmount: PromiseOrValue<BigNumberish>;
  swapCallData: PromiseOrValue<BytesLike>;
};